pyproject.toml
[project]
name = "agenticrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.24",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "mcp[cli]>=1.9.0",
    "notebook>=7.4.2",
    "python-dotenv>=1.1.0",
    "wikipedia>=1.4.0",
]
